Minor x86_64 (core2) kernel fixes.
Details:
- Fixed copy-and-paste bug whereby [scz]gemmtrsm_u_opt_d4x4 kernels
  for x86_64/core2 were calling the wrong reference code (l instead
  of u).
- Fixed some unused variables in x86_64/core2 dotaxpyv and dotxaxpyf
  kernels.
- Minor typecasting fix in testsuite/src/test_libblis.c.
- Makefile updates.

+++ testsuite/src/test_libblis.c
@@ -545,9 +545,9 @@ void libblis_test_output_params_struct( FILE* os, test_params_t* params )
 	// size is chosen automatically. We query the result of that automatic
 	// choice via sizeof(gint_t).
 	if ( BLIS_INT_TYPE_SIZE == 32 || BLIS_INT_TYPE_SIZE == 64 )
-		sprintf( intsize_str, "%d", BLIS_INT_TYPE_SIZE );
+		sprintf( intsize_str, "%d", ( int )BLIS_INT_TYPE_SIZE );
 	else
-		sprintf( intsize_str, "%d", sizeof(gint_t) * 8 );
+		sprintf( intsize_str, "%d", ( int )sizeof(gint_t) * 8 );
 
 	// Output some system parameters.
 	libblis_test_fprintf_c( os, "\n" );
